Fix error in publish.ps1
[skip ci]

+++ scripts/publish.ps1
@@ -17,8 +17,11 @@ ThrowOnError
 
 function Publish($package) {
     Set-Location $PSScriptRoot/../packages/$package
-    Remove-Item dist/*.test.js
-    Remove-Item dist/*.test.d.ts
+
+    if (Test-Path dist) {
+        Remove-Item dist/*.test.js
+        Remove-Item dist/*.test.d.ts
+    }
 
     yarn npm publish --access public
     ThrowOnError
